add exceptions.py, make excepthook a class
