+++ check50/internal.py
@@ -11,7 +11,8 @@ import traceback
 
 import lib50
 
-from . import _simple, __version__
+from . import _simple
+from .exceptions import Error
 
 #: Directory containing the check and its associated files
 check_dir = None
@@ -28,18 +29,6 @@ slug = None
 #: ``lib50`` config loader
 CONFIG_LOADER = lib50.config.Loader("check50")
 CONFIG_LOADER.scope("files", "include", "exclude", "require")
-
-
-class Error(Exception):
-    """Exception for internal check50 errors."""
-    pass
-
-
-class RemoteCheckError(Error):
-    """An exception for errors that happen in check50's remote operation."""
-    def __init__(self, remote_json):
-        super().__init__("check50 ran into an error while running checks! Please contact sysadmins@cs50.harvard.edu!")
-        self.payload = {"remote_json": remote_json}
 
 
 class Register:
@@ -190,72 +179,6 @@ def import_file(name, path):
     mod = importlib.util.module_from_spec(spec)
     spec.loader.exec_module(mod)
     return mod
-
-
-def _excepthook(cls, exc, tb):
-    """
-    check50's excepthook with configurable error output.
-
-    :ivar _excepthook.outputs: in which format errors should be returned (can be multiple)
-    :vartype _excepthook.outputs: tuple of strings, any of "json", "ansi", "html"
-    :ivar _excepthook.output_file: file to which the output should be redirected
-    :vartype _excepthook.output_file: str or pathlib.Path
-
-    See also: https://docs.python.org/3/library/sys.html#sys.excepthook
-    """
-
-    # If an error happened remotely, grab its traceback and message
-    if issubclass(cls, RemoteCheckError) and "error" in exc.payload["remote_json"]:
-        formatted_traceback = exc.payload["remote_json"]["error"]["traceback"]
-        show_traceback = exc.payload["remote_json"]["error"]["actions"]["show_traceback"]
-        message = exc.payload["remote_json"]["error"]["actions"]["message"]
-    # Otherwise, create the traceback and message from this error
-    else:
-        formatted_traceback = traceback.format_exception(cls, exc, tb)
-        show_traceback = False
-
-        if (issubclass(cls, Error) or issubclass(cls, lib50.Error)) and exc.args:
-            message = str(exc)
-        elif issubclass(cls, FileNotFoundError):
-            message = _("{} not found").format(exc.filename)
-        elif issubclass(cls, KeyboardInterrupt):
-            message = _("check cancelled")
-        elif not issubclass(cls, Exception):
-            # Class is some other BaseException, better just let it go
-            return
-        else:
-            show_traceback = True
-            message = _("Sorry, something is wrong! check50 ran into an error.\n" \
-                        "Please let CS50 know by emailing the error above to sysadmins@cs50.harvard.edu.")
-
-    # Output exception as json
-    if "json" in _excepthook.outputs:
-        ctxmanager = open(_excepthook.output_file, "w") if _excepthook.output_file else nullcontext(sys.stdout)
-        with ctxmanager as output_file:
-            json.dump({
-                "slug": slug,
-                "error": {
-                    "type": cls.__name__,
-                    "value": str(exc),
-                    "traceback": formatted_traceback,
-                    "actions": {
-                        "show_traceback": show_traceback,
-                        "message": message
-                    },
-                    "data" : exc.payload if hasattr(exc, "payload") else {}
-                },
-                "version": __version__
-            }, output_file, indent=4)
-            output_file.write("\n")
-
-    # Output exception to stderr
-    if "ansi" in _excepthook.outputs or "html" in _excepthook.outputs:
-        if show_traceback:
-            for line in formatted_traceback:
-                termcolor.cprint(line, end="", file=sys.stderr)
-        termcolor.cprint(message, "red", file=sys.stderr)
-
-    sys.exit(1)
 
 
 def _yes_no_prompt(prompt):
